Alle classes in verschillende bestanden gezet
